Implemented the LeafNode class. This means there was a change to the abstract Node class to solve the weird cyclic cycle of child calling parent.
# GitHub Issue Comment
## Update
I jumped into this to help with the LeafNode Class as my task #4 was a bit small. I am also altering the definition of split and insert to have a return of the new node created; specifically ((int,Data),Node). With the return, it is possible to split without an infinite loop of a child calling insert on the parent and a parent calling insert on the child.

## Note
The deleteKey method will require the parent to check if child needs to be merged after deletion.